requirements.txt
langchain==0.3.22
langchain_openai==0.3.12
langchain_community==0.3.7
chromadb==1.0.3
beautifulsoup4==4.13.3
requests==2.32.3
tiktoken==0.9.0
PyPDF2==3.0.1
gradio==4.24.0
python-dotenv==1.0.0